v1.5.4-hf3

+++ src/classes/actions/every/WrongChatAction.ts
@@ -2,6 +2,7 @@ import EveryMessageAction from './EveryMessageAction'
 import LinkedChatService from '../../db/services/linkedChat/LinkedChatService'
 import { EveryMessageOptions } from '../../../utils/values/types/action-options'
 import MessageUtils from '../../../utils/MessageUtils'
+import ChatService from '../../db/services/chat/ChatService'
 
 export default class WrongChatAction extends EveryMessageAction {
     constructor () {
@@ -20,6 +21,10 @@ export default class WrongChatAction extends EveryMessageAction {
             chatMember.status == 'left' ||
             chatMember.status == 'restricted'
 
+        if (!(isNotInChat || hasntLinkedChat)) {
+            return
+        }
+
         if (isNotInChat) {
             if (linkedChat) {
                 await LinkedChatService.remove(
@@ -28,13 +33,33 @@ export default class WrongChatAction extends EveryMessageAction {
                 )
             }
         }
-
-        if (isNotInChat || hasntLinkedChat) {
+        
+        const linkedChats = await LinkedChatService.getLinkedChats(id)
+        if (!linkedChats.length) {
             await MessageUtils.answerMessageFromResource(
                 ctx,
                 'text/every-action/wrong-chat.pug'
             )
             return true
         }
+        
+        const newChat = linkedChats[0]
+        const chatName = await ChatService.getName(newChat)
+        await LinkedChatService.set(
+            id,
+            newChat
+        )
+
+        await MessageUtils.answerMessageFromResource(
+            ctx,
+            'text/every-action/changed-linked.pug',
+            {
+                changeValues: {
+                    chatName
+                }
+            }
+        )
+
+        return true
     }
 }
